+++ src/pyramid_bimt/tests/test_views.py
@@ -6,6 +6,7 @@ from pyramid_basemodel import Session
 from pyramid_bimt import add_home_view
 from pyramid_bimt import configure
 from pyramid_bimt.testing import initTestingDB
+from pyramid.httpexceptions import HTTPNotFound
 
 import unittest
 import webtest
@@ -13,7 +14,10 @@ import webtest
 
 class TestLoginViewsFunctional(unittest.TestCase):
     def setUp(self):
-        self.config = testing.setUp()
+        settings = {
+            'bimt.app_title': 'BIMT',
+        }
+        self.config = testing.setUp(settings=settings)
         initTestingDB()
         configure(self.config)
         add_home_view(self.config)
@@ -47,8 +51,8 @@ class TestUserView(unittest.TestCase):
     def test_view_user(self):
         from pyramid_bimt.views import UserView
         from pyramid_bimt.models import User
-        request = testing.DummyRequest()
         context = User.get("admin@bar.com")
+        request = testing.DummyRequest()
         resp = UserView(context, request).view()
         self.assertEqual(resp["user"], context)
         self.assertEqual(list(resp["audit_log_entries"]), [])
@@ -57,7 +61,10 @@ class TestUserView(unittest.TestCase):
 
 class TestUserViewFunctional(unittest.TestCase):
     def setUp(self):
-        self.config = testing.setUp()
+        settings = {
+            'bimt.app_title': 'BIMT',
+        }
+        self.config = testing.setUp(settings=settings)
         initTestingDB()
         configure(self.config)
         app = self.config.make_wsgi_app()
@@ -70,7 +77,7 @@ class TestUserViewFunctional(unittest.TestCase):
     def test_view_user(self):
         self.config.testing_securitypolicy(
             userid='admin@bar.com', permissive=True)
-        resp = self.testapp.get('/users/admin@bar.com', status=200)
+        resp = self.testapp.get('/users/1', status=200)
         self.assertIn("admin@bar.com", resp.text)
 
     def test_view_users(self):
@@ -82,5 +89,64 @@ class TestUserViewFunctional(unittest.TestCase):
     def test_disable_enable_user(self):
         self.config.testing_securitypolicy(
             userid='admin@bar.com', permissive=True)
-        self.testapp.get('/users/admin@bar.com/disable', status=302)
-        self.testapp.get('/users/admin@bar.com/enable', status=302)
+        self.testapp.get('/users/1/disable', status=302)
+        self.testapp.get('/users/1/enable', status=302)
+
+    def test_actions_users(self):
+        self.config.testing_securitypolicy(
+            userid='admin@bar.com', permissive=True)
+        resp = self.testapp.get('/users', status=200)
+
+        # view user action
+        self.assertIn(
+            '<a href="http://localhost/users/1" title="View">',
+            resp.text
+        )
+
+        # edit user action
+        self.assertIn(
+            '<a href="http://localhost/users/1/edit"'
+            ' title="Edit">',
+            resp.text
+        )
+
+        # add user action
+        self.assertIn(
+            '<a class="pull-right" href="http://localhost/users/add">'
+            'Add User</a>',
+            resp.text
+        )
+
+
+class TestEditUserViewFunctional(unittest.TestCase):
+    def setUp(self):
+        settings = {
+            'bimt.app_title': 'BIMT',
+        }
+        self.config = testing.setUp(settings=settings)
+        initTestingDB()
+        configure(self.config)
+        app = self.config.make_wsgi_app()
+        self.testapp = webtest.TestApp(app)
+
+    def tearDown(self):
+        Session.remove()
+        testing.tearDown()
+
+    def test_add_user(self):
+        self.config.testing_securitypolicy(
+            userid='admin@bar.com', permissive=True)
+        resp = self.testapp.get('/users/add', status=200)
+        self.assertIn("<h1>Add User</h1>", resp.text)
+
+    def test_edit_user(self):
+        self.config.testing_securitypolicy(
+            userid='admin@bar.com', permissive=True)
+        resp = self.testapp.get('/users/1/edit', status=200)
+        self.assertIn("<h1>Edit User</h1>", resp.text)
+
+    def test_edit_no_user(self):
+        self.config.testing_securitypolicy(
+            userid='admin@bar.com', permissive=True)
+        self.assertRaises(
+            HTTPNotFound, self.testapp.get, '/users/123456789/edit')