+++ docs/CHANGELOG.rst
@@ -7,6 +7,9 @@ Changelog
 0.1.8 (unreleased)
 ------------------
 
+- Split ``views.py`` into ``views/`` sub-package.
+  [zupo]
+
 - Required options are ``mail.default_sender``, ``bimt.app_name``,
   ``bimt.app_title`` or application will fail at start. For example look
   at the ``development.ini``.
@@ -16,6 +19,7 @@ Changelog
   features are located in ``\users`` path. View/edit user options are in
   Options column for each member.
   [matejc]
+
 
 
 0.1.7 (2013-11-27)